Method 2
A Swift-based method of counting the total number of words in twokinds.
Notes: You need to add a folder conatining all webcomic pages you want to analyse, also, you'll have to change the filepath to said folder. the filepath already present is the path to my personal folder filled with twokinds stuff. It won't work for you.