Method 1
First, python-based, messy solution to find the wordcount for twokinds. It relies on another project, made by damishshah.
Notes: this method is less accurate, and takes a LONG time to run. It's a very messy folder, and, once more, you need to add your own folder containing the pages you want to analyse, as well as adding a link to that folder in the code itself.